# openapi.json
{"openapi":"3.0.0","info":{"version":"1.0.0","title":"TikTok Auth API","description":"Headless TikTok authentication API built with Better Auth and Hono"},"servers":[{"url":"https://4d575ba248ac.ngrok-free.app","description":"Current server"}],"tags":[{"name":"System","description":"System endpoints"},{"name":"Authentication","description":"OAuth authentication endpoints"},{"name":"User Profile","description":"User profile endpoints"},{"name":"TikTok Posting","description":"Content posting endpoints for TikTok"}],"components":{"schemas":{},"parameters":{}},"paths":{"/health":{"get":{"tags":["System"],"summary":"Health Check","description":"Check if the API is running","responses":{"200":{"description":"API is healthy","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["ok"]},"service":{"type":"string"}},"required":["status","service"]}}}}}}},"/oauth/tiktok":{"post":{"tags":["Authentication"],"summary":"Initiate TikTok OAuth","description":"Start the TikTok OAuth flow and get the authorization URL","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"provider":{"type":"string","enum":["tiktok"]}},"required":["provider"]}}}},"responses":{"200":{"description":"OAuth URL generated successfully","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string","format":"uri"},"redirect":{"type":"boolean"}},"required":["url","redirect"]}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Clear user session and logout","security":[{"cookieAuth":[]}],"responses":{"200":{"description":"Logout successful","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"}},"required":["success","message"]}}}}}}},"/api/auth/session":{"get":{"tags":["Authentication"],"summary":"Check Session","description":"Check if user is authenticated and get session info","responses":{"200":{"description":"Session info","content":{"application/json":{"schema":{"type":"object","properties":{"session":{"type":"object","nullable":true,"properties":{"id":{"type":"string"},"userId":{"type":"string"},"expiresAt":{"type":"string"}},"required":["id","userId","expiresAt"]},"user":{"type":"object","nullable":true,"properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"image":{"type":"string"}},"required":["id"]}},"required":["session","user"]}}}}}}},"/api/tiktok/profile":{"get":{"tags":["User Profile"],"summary":"Get TikTok Profile","description":"Fetch the authenticated user's TikTok profile information","security":[{"cookieAuth":[]}],"responses":{"200":{"description":"Profile fetched successfully","content":{"application/json":{"schema":{"type":"object","properties":{"profile":{"type":"object","properties":{"open_id":{"type":"string"},"union_id":{"type":"string"},"avatar_url":{"type":"string","format":"uri"},"display_name":{"type":"string"}},"required":["open_id","union_id","avatar_url","display_name"]},"sessionUser":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"image":{"type":"string"}},"required":["id"]}},"required":["profile","sessionUser"]}}}},"401":{"description":"Not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"details":{"nullable":true}},"required":["error"]}}}},"404":{"description":"No TikTok access token found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"details":{"nullable":true}},"required":["error"]}}}},"500":{"description":"Failed to fetch profile","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"},"details":{"nullable":true}},"required":["error"]}}}}}}},"/api/tiktok/creator-info":{"post":{"tags":["TikTok Posting"],"summary":"Query Creator Info","description":"Get the latest information about the authenticated TikTok creator","security":[{"cookieAuth":[]}],"responses":{"200":{"description":"Creator info retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"creator_avatar_url":{"type":"string"},"creator_username":{"type":"string"},"creator_nickname":{"type":"string"},"privacy_level_options":{"type":"array","items":{"type":"string"}},"comment_disabled":{"type":"boolean"},"duet_disabled":{"type":"boolean"},"stitch_disabled":{"type":"boolean"},"max_video_post_duration_sec":{"type":"number"}},"required":["creator_avatar_url","creator_username","creator_nickname","privacy_level_options","comment_disabled","duet_disabled","stitch_disabled","max_video_post_duration_sec"]}},"required":["data"]}}}},"401":{"description":"Not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/api/tiktok/post-video":{"post":{"tags":["TikTok Posting"],"summary":"Post Video to TikTok","description":"Initiate video upload to TikTok","security":[{"cookieAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"post_info":{"type":"object","properties":{"title":{"type":"string","maxLength":150,"description":"Video title (required, max 150 chars)"},"privacy_level":{"type":"string","enum":["PUBLIC_TO_EVERYONE","MUTUAL_FOLLOW_FRIENDS","SELF_ONLY"],"default":"PUBLIC_TO_EVERYONE","description":"Privacy setting for the video"},"disable_duet":{"type":"boolean","default":false,"description":"Disable duet feature"},"disable_comment":{"type":"boolean","default":false,"description":"Disable comments"},"disable_stitch":{"type":"boolean","default":false,"description":"Disable stitch feature"},"video_cover_timestamp_ms":{"type":"number","description":"Thumbnail timestamp in milliseconds"},"description":{"type":"string","description":"Video description/caption"},"auto_add_music":{"type":"boolean","default":false,"description":"Automatically add music"}},"required":["title"]},"source_info":{"type":"object","properties":{"source":{"type":"string","enum":["FILE_UPLOAD","PULL_FROM_URL"],"description":"FILE_UPLOAD: Upload file directly, PULL_FROM_URL: Provide video URL"},"video_url":{"type":"string","format":"uri","description":"HTTPS URL of video file (required for PULL_FROM_URL)"},"video_size":{"type":"number","description":"Size of video in bytes (required for FILE_UPLOAD)"},"chunk_size":{"type":"number","default":10000000,"description":"Chunk size in bytes for FILE_UPLOAD (default: 10MB)"},"total_chunk_count":{"type":"number","default":1,"description":"Number of chunks for FILE_UPLOAD (default: 1)"}},"required":["source"]}},"required":["post_info","source_info"]},"example":{"post_info":{"title":"My Amazing Video","privacy_level":"PUBLIC_TO_EVERYONE","disable_duet":false,"disable_comment":false,"disable_stitch":false,"description":"Check out this cool video! #tiktok #video"},"source_info":{"source":"PULL_FROM_URL","video_url":"https://example.com/video.mp4"}}}}},"responses":{"200":{"description":"Video post initiated successfully","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"publish_id":{"type":"string"},"upload_url":{"type":"string"}},"required":["publish_id"]}},"required":["data"]}}}},"401":{"description":"Not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/api/tiktok/post-photos":{"post":{"tags":["TikTok Posting"],"summary":"Post Photos to TikTok","description":"Initiate photo upload to TikTok","security":[{"cookieAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"post_info":{"type":"object","properties":{"title":{"type":"string","maxLength":150,"description":"Video title (required, max 150 chars)"},"privacy_level":{"type":"string","enum":["PUBLIC_TO_EVERYONE","MUTUAL_FOLLOW_FRIENDS","SELF_ONLY"],"default":"PUBLIC_TO_EVERYONE","description":"Privacy setting for the video"},"disable_duet":{"type":"boolean","default":false,"description":"Disable duet feature"},"disable_comment":{"type":"boolean","default":false,"description":"Disable comments"},"disable_stitch":{"type":"boolean","default":false,"description":"Disable stitch feature"},"video_cover_timestamp_ms":{"type":"number","description":"Thumbnail timestamp in milliseconds"},"description":{"type":"string","description":"Video description/caption"},"auto_add_music":{"type":"boolean","default":false,"description":"Automatically add music"}},"required":["title"]},"source_info":{"type":"object","properties":{"source":{"type":"string","enum":["FILE_UPLOAD","PULL_FROM_URL"]},"photo_cover_index":{"type":"number"},"photo_images":{"type":"array","items":{"type":"string"}}},"required":["source","photo_images"]},"post_mode":{"type":"string","enum":["DIRECT_POST"]},"media_type":{"type":"string","enum":["PHOTO"]}},"required":["post_info","source_info","post_mode","media_type"]}}}},"responses":{"200":{"description":"Photo post initiated successfully","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"publish_id":{"type":"string"}},"required":["publish_id"]}},"required":["data"]}}}},"401":{"description":"Not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/api/tiktok/post-status":{"post":{"tags":["TikTok Posting"],"summary":"Check Post Status","description":"Check the status of a TikTok post","security":[{"cookieAuth":[]}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"publish_id":{"type":"string"}},"required":["publish_id"]}}}},"responses":{"200":{"description":"Post status retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"status":{"type":"string"},"fail_reason":{"type":"string"},"publicaly_available_post_id":{"type":"array","items":{"type":"string"}},"uploaded_bytes":{"type":"number"}},"required":["status"]}},"required":["data"]}}}},"401":{"description":"Not authenticated","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string"}},"required":["error"]}}}}}}},"/api/tiktok/upload-video":{"put":{"tags":["TikTok Posting"],"summary":"Upload Video File","description":"Upload video file to TikTok (used with FILE_UPLOAD source)","security":[{"cookieAuth":[]}],"parameters":[{"schema":{"type":"string"},"required":true,"name":"upload_url","in":"query"}],"requestBody":{"content":{"video/mp4":{"schema":{"nullable":true}}}},"responses":{"200":{"description":"Video uploaded successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"}},"required":["success"]}}}}}}}}}